fix(macOS): :bug: Improve macOS experience
This change primarily affects the server. It now checks we last sent data and if we did we push it to the stack. Otherwise it is ignored. This tries to mitigate the charging issue ( We put the headphones to charge it sends a message without requiring a command to be sent)